+++ designer/src/smart-editor/editor.ts
@@ -1,13 +1,13 @@
 import { Step } from '../definition';
-import { DesignerApi } from '../api/designer-api';
 import { EditorRenderer } from '../api/editor-renderer';
 import { EditorView } from './editor-view';
 import { GlobalEditorProvider, StepEditorProvider } from '../designer-configuration';
+import { EditorApi } from '../api';
 
 export class Editor {
 	public static create(
 		parent: HTMLElement,
-		api: DesignerApi,
+		api: EditorApi,
 		stepEditorClassName: string,
 		stepEditorProvider: StepEditorProvider,
 		globalEditorClassName: string,
@@ -16,21 +16,22 @@ export class Editor {
 		const view = EditorView.create(parent);
 
 		function render(step: Step | null) {
+			const definition = api.getDefinition();
 			let content: HTMLElement;
 			let className: string;
 			if (step) {
-				const stepContext = api.editor.createStepEditorContext(step.id);
-				content = stepEditorProvider(step, stepContext);
+				const stepContext = api.createStepEditorContext(step.id);
+				content = stepEditorProvider(step, stepContext, definition);
 				className = stepEditorClassName;
 			} else {
-				const globalContext = api.editor.createGlobalEditorContext();
-				content = globalEditorProvider(api.editor.getDefinition(), globalContext);
+				const globalContext = api.createGlobalEditorContext();
+				content = globalEditorProvider(definition, globalContext);
 				className = globalEditorClassName;
 			}
 			view.setContent(content, className);
 		}
 
-		const renderer = api.editor.runRenderer(step => render(step));
+		const renderer = api.runRenderer(step => render(step));
 		return new Editor(view, renderer);
 	}