+++ react/src/StepEditorWrapper.spec.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render, screen } from '@testing-library/react';
-import { Step, StepEditorContext } from 'sequential-workflow-designer';
-import { StepEditorWrapperContext } from './StepEditorWrapper';
+import { Definition, Step, StepEditorContext } from 'sequential-workflow-designer';
+import { StepEditorWrapperContext, useStepEditor } from './StepEditorWrapper';
 
 describe('StepEditorWrapper', () => {
 	const step: Step = {
@@ -10,6 +10,10 @@ describe('StepEditorWrapper', () => {
 		name: 'z',
 		properties: {},
 		type: 'test'
+	};
+	const definition: Definition = {
+		sequence: [step],
+		properties: {}
 	};
 	const context: StepEditorContext = {
 		notifyNameChanged() {
@@ -24,12 +28,28 @@ describe('StepEditorWrapper', () => {
 	};
 
 	it('renders child correctly', () => {
+		function TestHook() {
+			const editor = useStepEditor();
+
+			expect(editor.id).toBe('y');
+			expect(editor.type).toBe('test');
+			expect(editor.componentType).toBe('x');
+			expect(editor.name).toBe('z');
+			expect(editor.step).toBe(step);
+			expect(editor.properties).toBe(step.properties);
+			expect(editor.definition).toBe(definition);
+
+			return <hr data-testid="hook" />;
+		}
+
 		render(
-			<StepEditorWrapperContext step={step} context={context}>
+			<StepEditorWrapperContext step={step} definition={definition} context={context}>
 				<div data-testid="child" />
+				<TestHook />
 			</StepEditorWrapperContext>
 		);
-		const element = screen.getByTestId('child');
-		expect(element.tagName).toBe('DIV');
+
+		expect(screen.getByTestId('child').tagName).toBe('DIV');
+		expect(screen.getByTestId('hook').tagName).toBe('HR');
 	});
 });