+++ react/src/StepEditorWrapper.tsx
@@ -1,24 +1,25 @@
-import { Context, createContext, useContext, useState } from 'react';
-import { StepEditorContext, PropertyValue, Step } from 'sequential-workflow-designer';
+import { Context, createContext, useContext, useState, ReactNode } from 'react';
+import { StepEditorContext, PropertyValue, Step, Definition } from 'sequential-workflow-designer';
 
 declare global {
 	interface Window {
-		sqdStepEditorContext?: Context<StepEditorWrapper<Step> | null>;
+		sqdStepEditorContext?: Context<StepEditorWrapper | null>;
 	}
 }
 
 if (!window.sqdStepEditorContext) {
-	window.sqdStepEditorContext = createContext<StepEditorWrapper<Step> | null>(null);
+	window.sqdStepEditorContext = createContext<StepEditorWrapper | null>(null);
 }
 const stepEditorContext = window.sqdStepEditorContext;
 
-export interface StepEditorWrapper<TStep extends Step> {
+export interface StepEditorWrapper<TStep extends Step = Step, TDefinition extends Definition = Definition> {
 	readonly id: string;
 	readonly type: TStep['type'];
 	readonly componentType: TStep['componentType'];
 	readonly name: string;
 	readonly properties: TStep['properties'];
 	readonly step: TStep;
+	readonly definition: TDefinition;
 
 	setName(name: string): void;
 	setProperty(name: keyof TStep['properties'], value: TStep['properties'][typeof name]): void;
@@ -26,18 +27,28 @@ export interface StepEditorWrapper<TStep extends Step> {
 	notifyChildrenChanged(): void;
 }
 
-export function useStepEditor<TStep extends Step = Step>(): StepEditorWrapper<TStep> {
+export function useStepEditor<TStep extends Step = Step, TDefinition extends Definition = Definition>(): StepEditorWrapper<
+	TStep,
+	TDefinition
+> {
 	const wrapper = useContext(stepEditorContext);
 	if (!wrapper) {
 		throw new Error('Cannot find step editor context');
 	}
-	return wrapper as unknown as StepEditorWrapper<TStep>;
+	return wrapper as unknown as StepEditorWrapper<TStep, TDefinition>;
 }
 
-export function StepEditorWrapperContext(props: { children: JSX.Element; step: Step; context: StepEditorContext }) {
+export interface StepEditorWrapperContextProps {
+	children: ReactNode;
+	step: Step;
+	definition: Definition;
+	context: StepEditorContext;
+}
+
+export function StepEditorWrapperContext(props: StepEditorWrapperContextProps) {
 	const [wrapper, setWrapper] = useState(() => createWrapper());
 
-	function createWrapper(): StepEditorWrapper<Step> {
+	function createWrapper(): StepEditorWrapper {
 		return {
 			id: props.step.id,
 			type: props.step.type,
@@ -45,6 +56,7 @@ export function StepEditorWrapperContext(props: { children: JSX.Element; step: S
 			name: props.step.name,
 			properties: props.step.properties,
 			step: props.step,
+			definition: props.definition,
 			setName,
 			setProperty,
 			notifyPropertiesChanged,